Removed search old command and moved api2 to cli
